requirements.txt
apache-airflow[postgres]==2.5.1
psycopg2-binary==2.9.5